doc: improve documentation
- rename LWFSM as LwFSM
- fix doxygen issues leading to undocumented items
- use @ instead of \ as markup

+++ examples/stm32/lwfsm_demo/App/lwfsm_app.c
@@ -1,6 +1,6 @@
 /**
- * \file            lwfsm_app_template.c
- * \brief           LwFSM application file
+ * @file            lwfsm_app_template.c
+ * @brief           LwFSM application file
  */
 
 /*
@@ -35,6 +35,7 @@
 /* Add your own includes here */
 #include "main.h"
 #include "lwfsm.h"
+#include "lwfsm_app.h"
 
 /* declare your FSM state C function prototypes here */
 static uint32_t prv_initial_state(void * opaque_ctxt);
@@ -68,7 +69,7 @@ LWFSM_TABLE_END(my_state_machine)
 /* FSM task entry point */
 void myAppTaskM1(void * p_ctxt)
 {
-  lwfsm_status_t ret = lwfsm_init_state_machine(&my_lwfsm_ctxt, my_state_machine, my_states_names, INIT_STATE);
+  lwfsm_status_t ret = lwfsm_init_state_machine(&my_lwfsm_ctxt, my_state_machine, my_states_names, (uint32_t)INIT_STATE);
 
   if (LWFSM_OK == ret)
   {
